muda tipos para float

+++ src/main.hpp
@@ -2,12 +2,12 @@
 #define __MAIN_H__
 
 int sizeInput();
-void matrixInput(double **m, int size);
-void vectorInput(double *vetor, int size);
-void printMatrix(double **m, int size);
-void concatenateMatrix(double **m, double *vetor, int size);
-void solve(double **m, int size);
-void updateRow(double **m, int row, double factor, int size, int iteration);
+void matrixInput(float **m, int size);
+void vectorInput(float *vetor, int size);
+void printMatrix(float **m, int size);
+void concatenateMatrix(float **m, float *vetor, int size);
+void solve(float **m, int size);
+void updateRow(float **m, int row, float factor, int size, int iteration);
 int main();
 
 struct fraction {
